Documents/TransL/backend/requirements.txt
fastapi==0.135.2
uvicorn==0.42.0
starlette==1.0.0

pymongo==4.17.0
dnspython==2.8.0

python-dotenv==1.2.2

python-jose==3.5.0
PyJWT==2.12.1
rsa==4.9.1
ecdsa==0.19.1
pyasn1==0.6.3

passlib==1.7.4
bcrypt==4.0.1

python-multipart==0.0.22
email-validator==2.3.0

cryptography==46.0.5
cffi==2.0.0
pycparser==3.0

click==8.3.1
anyio==4.13.0
h11==0.16.0
idna==3.11
six==1.17.0
typing_extensions==4.15.0